feat(index): improve public API
The config builder class needs to be exported so that it can be used as
a parameter, but I don't want it to be part of the public API. For this
reason, specific things are now exported from the index file. If
something is exported, but not from the index file, it is not considered
to be part of the public API. This means that breaking changes to those
things would not warrant a major version increment.